Rj/services

Approved-by: Roch Bajracharya

+++ src/app/components/leave_screen/deleteAlert.tsx
@@ -1,6 +1,7 @@
 import React, { useContext, useState } from 'react';
-import { View, TouchableOpacity } from 'react-native';
-import Dialog from 'react-native-dialog';
+import { View, TouchableOpacity, Text } from 'react-native';
+import normalize from 'react-native-normalize';
+import { ConfirmDialog } from 'react-native-simple-dialogs';
 import colors from '../../../assets/colors';
 import { deleteAlertStyle as style } from '../../../assets/styles';
 import { AppIcon, snackBarMessage } from '../../common';
@@ -8,7 +9,6 @@ import { dataType } from '../../interface';
 import { RequestContext, TimeLogContext } from '../../reducer';
 import { deleteRequest, cancelLeave } from '../../services';
 import { deleteTimeLog } from '../../services/timeLogService';
-import { getUser } from '../../utils';
 
 const DeleteAlert = ({
   item,
@@ -33,7 +33,6 @@ const DeleteAlert = ({
   const [loading, setLoading] = useState(false);
 
   const onDelete = async () => {
-    const user = await getUser();
     setLoading(true);
     if (other) {
       cancelLeave(item.id)
@@ -83,35 +82,36 @@ const DeleteAlert = ({
           size={23}
         />
       </TouchableOpacity>
-      <Dialog.Container
+      <ConfirmDialog
         visible={showAlert}
-        contentStyle={style.dialogContainer}
+        onTouchOutside={() => setShowAlert(false)}
+        contentStyle={[style.innercontent, { marginBottom: normalize(-30) }]}
+        titleStyle={style.text1}
+        positiveButton={{
+          titleStyle: style.delete,
+          title: 'DELETE',
+          onPress: () => {
+            timelog ? onTimeLogDelete() : onDelete();
+            hide();
+          },
+        }}
+        negativeButton={{
+          titleStyle: style.cancel,
+          title: 'CANCEl',
+          onPress: () => hide(),
+        }}
       >
         <View style={style.container}>
           <AppIcon name="alert" color={colors.tomato} size={30} />
           <View style={style.main}>
-            <Dialog.Title style={style.text1}>
+            <Text style={style.text1}>
               {other ? 'Cancel' : 'Delete'} the{' '}
               {edittimelog ? 'task ' : timelog ? 'timelog' : 'request'} ?
-            </Dialog.Title>
-            <Dialog.Title style={style.text2}>This cant be undone</Dialog.Title>
+            </Text>
+            <Text style={style.text2}>This can't be undone</Text>
           </View>
         </View>
-        <View style={style.buttons}>
-          <Dialog.Button
-            label={other ? 'NO' : 'CANCEL'}
-            onPress={hide}
-            style={style.cancel}
-          />
-          <Dialog.Button
-            label={other ? 'YES' : 'DELETE'}
-            onPress={() => {
-              timelog ? onTimeLogDelete() : onDelete();
-            }}
-            style={style.delete}
-          />
-        </View>
-      </Dialog.Container>
+      </ConfirmDialog>
     </>
   );
 };
